REFACTOR: Move model/analysis documents to resp. top-level folders.
The top-level `document` folder has been removed.
