--- packages/frontend/src/document/index.ts
@@ -1,3 +1,0 @@
-export * from "./types.ts";
-export * from "./model_document_editor.tsx";
-export * from "./analysis_document_editor.tsx";
