+++ packages/frontend/src/model/index.ts
@@ -1,4 +1,7 @@
 export * from "./types";
 export * from "./context";
+export * from "./document";
+
 export * from "./object_cell_editor";
 export * from "./morphism_cell_editor";
+export * from "./model_editor";
