+++ packages/frontend/src/model/theory_selector.css
